added pause function-refactor Scene

+++ src/main/java/org/ca/bcit/comp2522/Bubble.java
@@ -85,9 +85,9 @@ public class Bubble extends Sprite implements Collidable{
     PVector newVelocity1 = new PVector(-velocity.x, velocity.y < 0 ? velocity.y : -velocity.y);
     PVector newVelocity2 = new PVector(velocity.x, velocity.y < 0 ? velocity.y : -velocity.y);
     Bubble bubble1 = new Bubble(new PVector(position.x, position.y), new PVector(-1, -1),
-        newSize, speed, color, window, newVelocity1);
+        newSize, (float) speed, color, window, newVelocity1);
     Bubble bubble2 = new Bubble(new PVector(position.x, position.y), new PVector(1, -1),
-        newSize, speed, color, window, newVelocity2);
+        newSize, (float) speed, color, window, newVelocity2);
     newBubbles.add(bubble1);
     newBubbles.add(bubble2);
     return newBubbles;
