+++ src/main/java/org/ca/bcit/comp2522/ShootLine.java
@@ -41,10 +41,10 @@ public class ShootLine extends Sprite {
     PVector currRed = new PVector(x - thickness /2, y);
     PVector prevYellow = new PVector(x - thickness /2, y);
     PVector prevRed = new PVector(x + thickness /2, y);
-    float maxIncrements = increments - this.getPosition().y / this.speed;
+    float maxIncrements = (float) (increments - this.getPosition().y / this.speed);
     for(float i = 0; i < maxIncrements - speed; i++) {
-      currYellow.y = prevYellow.y - this.speed;
-      currRed.y = prevRed.y - this.speed;
+      currYellow.y = (float) (prevYellow.y - this.speed);
+      currRed.y = (float) (prevRed.y - this.speed);
       window.stroke(255, 204, 0);
       window.line(currYellow.x, currYellow.y, prevYellow.x, prevYellow.y);
       window.stroke(255, 0, 0);
@@ -69,7 +69,7 @@ public class ShootLine extends Sprite {
    */
   public void update(GameWindow window) {
     if(this.position.y > 100 + thickness / 2) {
-      this.position.y = this.position.y - speed;
+      this.position.y = (float) (this.position.y - speed);
     }
   }
